+++ tests/dyn.vars/src/subproject/nested_folder2/indexUtil.js
@@ -1,0 +1,9 @@
+//@ts-check
+// import { B as _B1 } from './util1';
+
+export default {
+    a: 2
+};
+// export const B = _B1;
+
+// console.log(kkk)