+++ tests/dyn.vars/src/subproject/nested_folder2/util1.js
@@ -1,0 +1,19 @@
+//@ts-check
+// export const a = 1;
+
+/// have some bug for dynimports (// FIXME):
+// module.exports.a = () => {
+    
+// }
+
+// const a = 1;
+// export {
+//     a
+// }
+
+const b = 1;
+const c = 8;
+export {
+    b as B,
+    c
+}